sysctl: remove redundant code
The xstrdup function calls strdup internally,
and strdup will apply for memory internally
and needs to be released externally. The
function calls xstrdup(name) twice, but only
releases the memory requested by the second
call. The first call of xstrdup(name) is
meaningless, so it is recommended to delete it.

signed-off-by: zhoujie <zhoujie133@huawei.com>