top: modifying the behavior of the '!' toggle function
Previously, each press of `!' led to the combining of an additional
CPU. However, we have now updated this feature to double the number of combined
CPUs with each press.

The updated example now reflects this change: with the first press of !', two
CPUs are combined and displayed as 0-1, 2-3, ...', instead of the previous
format. Additional presses of `!' continue to double the number of combined
CPUs.

[ jw note: the commit message 1st line was amended to ]
[ add the top program name and remove capitalization. ]
[ the patch itself was adjusted to avoid a fatal po4a ]
[ man document error encountered during clean builds. ]

Signed-off-by: Vishal Chourasia <vishalc@linux.ibm.com>
Signed-off-by: Jim Warner <james.warner@comcast.net>