pmap: Reset totals between procs with -X or -XX
When pmap used with -X or -XX option with multiple pids, the
totals were not reset between the pids, meaning the first total
was for pid1, but second total was for pid1+pid2.

These running totals are now reset when they are printed, meaning
each total is for it's own pid only.

References:
 issue #298

Signed-off-by: Craig Small <csmall@dropbear.xyz>