library: include the 'docker' containers in <pids> api
This commit makes a task's 'docker' container id (both
full and abbreviated) available within the <pids> API.

It is patterned after that 'lxc' container logic which
tracks ids already seen so as to avoid repeated malloc
plus free calls. It should be noted, however, that the
'docker' ids do not proliferate like those others may.

In implementing the addition, the 'readproc/task' code
had to be tweaked a little. The code for PROC_FILL_LXC
and PROC_FILL_DOCKER now share a buffer containing the
cgroup contents so as to avoid a second file2str call.

[ yes, each container extraction routine may corrupt ]
[ that buffer. and yes, a caller may request both of ]
[ those enums (LXNAME and DOCKER_ID) simultaneously. ]
[ but, a task can be in only one container at a time ]
[ so the buffer would get whacked at most just once. ]

Signed-off-by: Jim Warner <james.warner@comcast.net>