w: Fix check for empty username
In the referenced commit the check for an empty ut_user was incorrectly
checking if the structure existed.

From:
-                       if (u->ut_type != USER_PROCESS)
-                               continue;
-                       if (*u->ut_user)
-                               showinfo(

To:
+               if (u->ut_type != USER_PROCESS || (NULL == u->ut_user))

The previous code somewhat opaquely was checking if ut_user was "\0"
and if it was not show the information.

This commit reverts the function of the previous code, but does it in
an explicit way to check '\0' == ut_user[0] so there is no confusion
about pointers and characters.

References:
 commit 1c61ac2c6180f241a255967ca0bba4e944643440