slabtop: Fix test and documentation
In the referenced commit, the slab sizes use the scale_size()
function to print the values. This is to enable the --human
flag so it auto-scales.

As a side-effect, the default printout went from %.2fK to just %d
which is what the other values used and makes the output cleaner.

Unfortunately, two items were missed.
1) The testsuite did not have the regex updated. As slabtop tests are
generally missed due to permisions to /proc/slabtop this often doesn't
show up, except sometimes on CI infrastructure.
2) The man page was silent what the units were used for the sizes when
the --human flag was absent.

References:
 commit c4f33bc302c0589261a69a030935e3cb437e9daf

Signed-off-by: Craig Small <csmall@dropbear.xyz>