ps: correct truncation using '%a' AIX format specifier
Thanks to that issue referenced below, an earlier oops
contained in a commit which purported to make some aix
parsing logic more robust was revealed. In that commit
one code line was relocated outside an iterative path.
That was done so as to catch extraneous invalid chars.

As a consequence, 'items' would be overstated and that
produced a null/empty format node always created last.
It's that node which received the balance of available
buffer space, not the desired '%a' (args/COMMAND) col.

[ contrary to the issue's claim, truncation occurred ]
[ at 27 characters not 15. such a default args width ]
[ is used when any other field appears to the right. ]

While we're at it, we'll address yet another edge case
bug found even in a version 3 ps. When one begins some
aix format string with a random character which is not
preceded by a '%' specifier, such a character would be
accepted then output along with the header & the data.

Reference(s):
https://gitlab.com/procps-ng/procps/-/issues/316
. Apr, 2022 - aix parsing made more robust
commit 4fbf8d22a99445bc7b37c6d36ed9e045821439e3

Signed-off-by: Jim Warner <james.warner@comcast.net>