library: Add procps_uptime_snprint
Uses the same format as the current procps_uptime_sprint*
functions.

Uptime is given, not assumed. This is so if, for example,
container uptime is given then the function will work.

Non-pretty format uses same logic as pretty format for
days and hours to be consistent.

Uses a given string buffer like snprintf() instead of a
static internal buffer.

Partly solves container uptime issue #300

Signed-off-by: Craig Small <csmall@dropbear.xyz>