library: Add procps_container_uptime
This function will put the container's uptime in seconds into the
given variable. It is based upon the start time of process 1
which, for most types of containers, is the same as when the
container starts.

I'm sure there will be some odd form of container where this
is not the case :/

Signed-off-by: Craig Small <csmall@dropbear.xyz>