vmstat: Refactor single printfs as table of column widths and headers
This refactors the header-and-column-printing code to use a table
of widths instead of huge printf strings (exception: the first line
of the header that group the columns).

It is preparatory work to be able to behave better when output values
overflow their columns.

This makes no functional change to vmstat.

Signed-off-by: Craig Small <csmall@dropbear.xyz>