top: prevent rcfile inspect/other-filter error message
It is almost as if stdio fgets() used to always insert
the '\0' byte, even at EOF, but now no longer does so.

[ either that or testing of 3.3.16 was badly lacking ]

In any case, this single line insertion will ensure an
empty buffer even when fgets() returns a NULL for EOF.
That's a requirement if we're to avoid this error msg.

Reference(s):
. Jun, 2018 - 'other filters' saved in rcfile
commit 35e9269bbe0a7c3f3b2ef999a281e447dc7f4ab4

Signed-off-by: Jim Warner <james.warner@comcast.net>