library: Better escape functions for UTF-8
Changed so all non-printables show up as '?' instead of
'?', '.' or '@' depending (and not consistently for the
same character).

esc_all() moved from using a static table to using isprint()
If we can't print it, we '.' it

esc_ctl() has been greatly simplified and is more robust.
This is largely due to the work @zilka did in !168
(but needed work to merge in the new layout) around
u8charlen() where the function checks for both the first
byte and subsequent bytes, e.g. For UTF8-3 you can't just check
The first byte is E0-EF because EC BF is valid but ED BF is not,
same with UTF8-4 and F0 BF vs F4 BF; such is the fun that is Unicode.

esc_ctl() now follows the RFC3629 UTF-8 byte sequences and
has a reasonable parser now. I'm sure something will break it
given Unicode's complexity and the fact we cannot assume we
get valid data because it depends on whatever you call your
program or dump in prctl().

References:
 https://datatracker.ietf.org/doc/html/rfc3629#section-4
 https://www.unicode.org/versions/latest/
 !168

Signed-off-by: Craig Small <csmall@dropbear.xyz>