testsuite: Test the library escape routines
This test program directly includes the library's escape.c
as we are testing internal static functions.

The previous commit I mentioned that the escape parser
will probably break with some future Unicode horror,
now that horror can be put into the test strings!

check_size_strlen() check_size_negative() are largely
lifted from the comments in !168 written by @zilka
check_size_ascii is a variation on the above, all
ASCII characters should return 1.

check_esc_ctl() has an array of input and output strings,
put your Unicode horrors in cell 0 and the expected clean
output in cell 1. It has some of the ANSI color sequences
as well as C1 controls that the previous commit fixed
plus some malformed strings from Markus Kuhn's page.

If there is a program name that ps or top chokes on,
the array in check_esc_ctl() is the place to put that
name in.

References:
 commit 49f92f690cf18a432b33943c671f560f90dae204
 https://www.cl.cam.ac.uk/~mgk25/ucs/examples/UTF-8-test.txt
 !168

Signed-off-by: Craig Small <csmall@dropbear.xyz>