library: make the better UTF-8 escape functions better
In the commit referenced below, that escape module was
vastly simplified and made more robust in the process.

However, there is one overlooked area that will always
present display problems. That's the Private Use Area.

And, while there are actually 3 PUAs, this commit will
focus on the main range in the BMP (U+E000 to U+F8FF).

[ along the way, we'll tweak a few whitespace issues ]

Reference(s):
. May, 2024 - better escape functions
commit 73d332e8f174d51ddf44a14573bac623601e8c00

Signed-off-by: Jim Warner <james.warner@comcast.net>