ps: Another go at AIX formats
Previous commits around AIX formats fixed some issues but meant that some
of the useful features of that format were missing.

The problem is that AIX uses %* and standard formats use %* too for
%cpu and %mem.

The trick is to use AIX parser when there is a % but not when its
for the "normal" fields.

%mem is easy as there is no %m AIX equivalent so "%m*" means
don't use AIX parser.

%cpu is harder. "%cpu" "%cpu pid" or "%cpu,pid" are to be
parsed normally, but "%cpuXpid" is AIX "%c Xpid".

I think I've caught all the odd cases, but I'm sure there is some
strange combination I have missed.

References:
 #323
 commit 8cb646bdfc6e37b221f9616df96c6c116e6def28
Signed-off-by: Craig Small <csmall@dropbear.xyz>