top: suppress compatibility warning for SYS_RCDEFAULTS
Way back when the ultimate solution to the top startup
defaults was introduced, in that commit shown below, a
detail was overlooked that might raise some questions.

It was suggested that the easiest way to establish the
/etc/topdefaultrc file was for a sysadmin to customize
top to their liking, save the rcfile with 'W' and then
just move it to that /etc directory as 'topdefaultrc'.

That was fine then, but over time the ID in the rcfile
(RCF_VERSION_ID) would change. This then would produce
a compatibility warning message if a user later issued
a 'W' command to save their own rcfile the first time.

Now, with this patch that warning is being suppressed.

Reference(s):
. Dec, 2017 - introduced default system rcfile
commit 3e6a208ae501194fdb39d5f259e327c087dc8c84

Signed-off-by: Jim Warner <james.warner@comcast.net>