top: correct the irix mode calculation for maximum cpu
When an opensuse patch for large numbers of processors
was borrowed, in that commit shown below, a small oops
was also ported. This patch will correct that mistake.

However, as noted in the commit message cited below, a
user will see '?' in that %CPU column if ever Cpu_pmax
was set at the 2 highest levels in this chunk of code.

That's because this field can display at most 5 chars.

[ but thanks to John Fingerhut for raising the issue ]

Reference)s):
https://gitlab.com/procps-ng/procps/-/issues/338
. Apr, 2022 - imported opensuse patch
commit 7279cd8e4d21f8fab62e34e4695b46e86cba2442

Signed-off-by: Jim Warner <james.warner@comcast.net>