Library: Recover from meminfo seek() in LXC
Running procps tools under LXC may fail with the message:

"Unable to create meminfo structure"

This is due to a lseek() failing with a ESPIPE error.
It doesn't appear to impact all LXC installations and it's
unclear why some are impacted and some not.

The fix is to check for that error and reopen the file
if required.

References:
  https://bugs.debian.org/1072831

Signed-off-by: Craig Small <csmall@dropbear.xyz>