library: Provide sigabbrev_np() replacement
The original idea of using signal_number_to_name() was ok when the
calling function was in ps, but problematic in the library because:
* There needed to be a libtool and non-libtool built of local/signals.c
* The function handled RT signals, but the calling function in the
  library does this too.
* It needed to be an array of signal,name pairs because signals.c
  is two-way while we only need num->name in the library.
* signal_number_to_name is called by signal handlers so less library
  calls the better.

If sigabbrev_np() is available in libc, then the library should use
that. This commit provides a back-up version of that function which
is only used by libc's that can't provide.

Greatly inspired by Peter Hutterer's work-around with some clever
stringizing and concatenation in the preprocessor for same issue
in Pipewire.

References:
 commit 7f56a9aea1456568202f8cb240a76cebf351da5f
 https://gitlab.freedesktop.org/pipewire/pipewire/-/commit/731f45ed5012ddfd2a34a272359eab1062f13e7b
 https://gcc.gnu.org/onlinedocs/cpp/Stringizing.html

Signed-off-by: Craig Small <csmall@dropbear.xyz>