docs: watch.8 --chgexit only on visible changes
watch only exits on changes that are visible. This is a by-product
of how changes are detected which is in the ncurses idea of windows.

The limits of change detection is also why multicolours can't be
implemented, so a large re-work of how the output is stored and
displayed may fix both.

For now a note in the manpage is all we can do.

References:
 https://bugs.debian.org/729569

Signed-off-by: Craig Small <csmall@dropbear.xyz>