docs: add a caveat for one function in that <pids> API
A recent issue, now closed, was raised by a brave soul
who successfully took on and exploited the <pids> API.

That program contained an iterative procps_pids_select
involving the UTILIZATION_C plus TICS_ALL_DELTA items.

Also included in the loop was a procps_pids_reset call
which may have been present just to produce some extra
loading so as to increase the UTILIZATION_C value. But
I began to wonder if its presence was maybe a mistake.

After all, nowhere in the documentation was it clearly
stated that even with an iterative process the user is
not required to clean/reset the stacks with each pass.

So this commit will clarify the conditions under which
a procps_pids_reset call was really needed. Of course,
if such advice is ignored, and extra reset calls made,
no real harm is done. Just some cpu cycles are wasted.

Reference(s):
https://gitlab.com/procps-ng/procps/-/issues/349

Signed-off-by: Jim Warner <james.warner@comcast.net>