ps: Zombies get <defunct> on command name again
Version 3.3.17 would append <defunct> to the command line and command name
if the process was a zombie (state Z).
As part of the escape function fixes, the command name lost this.

The feature is now restored, in addition pr_args() and pr_comm() had a
lot of duplicate code (and would have had more for this fix) so they
now only do the selection part and call a common function.

The sitation now in the library is command names do not get <defunct>
appended, but command lines do.  It's not really the library's job
to change the presentation of the command line (only provide the status
field) but this change now makes procps behave as previous versions.

There are now checks for both linux and bsd formatting to check the
command line/name are correctly marked.

References:
 #355
 commit 7b0fc19e9d28380dc9790615b93bc3653d6d686e

Signed-off-by: Craig Small <csmall@dropbear.xyz>