pgrep: Include sys/syscall.h if making pidwait
sys/syscall.h would only be included if pidwait was made
and we found pidfd_open() The previous commit fixed the
finding part, but in fact we want sys/syscall.h either
way because syscall() is defined there too.

Most of the time the header is included by other headers
but adding it explicitly means if that header is removed or
changed it still works.

Signed-off-by: Craig Small <csmall@dropbear.xyz>