build-sys: Be specific about translation files
At dist time, do not update the po files

If the po-man/*.po files are updated make distcheck returns an
error because they are updated and found in the _build directory.

These files should only be updated when we are fiddling around
with translations, e.g. updating them *before* a dist

update-po - Do not do translations
Only update the po and pot files for update-po