pgrep.1: Add note about process ages
pgrep -o and -n print the oldest or newest matched process. If
processes start at least one clock tick apart, then the match is
always correct, otherwise its the first and last found process
that started within that clock tick (usually 1/100 second) because
thats the resolution of starttime in /proc/pid/stat

It's generally masked because the oldest process has the lowest
pid and the newest process has the largest... except when the
pids wrap around.

Added a note to pgrep.1 mentioning this limitation.

References:
 #376