w: Add terminal mode
Additional option for w -t/--terminal

Instead of running through either the systemd sessions or utmp
entries this option tries to find all ttys that have users on them.
It currently looks at /dev/tty* and /dev/pts/* but could be
extended.

getty processes are assumed to have a PPID or 0 or 1, to be
excluded from user processes.

The format and most flags are the same, the order is whatever
opendir() decides, which is random-ish.

Some noteable differences:
 * session leaders for utmp systems will be one process below
   which is the first process running on that device, this
   changes the -p option
 * Remote hosts are based on the session, not device. So
   a user that ssh's and runs screen will have the ssh session
   and the screens showing a remote IP. Almost no tools
   differentiate between the session running screen and its
   screens so its not real big difference.
 * From/IP won't work for utmp systems

There could be work to get utmp to get session leaders and
remotes but it's a lot of trawling through utmpx to get it.

References:
 #375

Signed-off-by: Craig Small <csmall@dropbear.xyz>