w: refactor the terminal mode
The process stack is both able to be sorted and has tty devices,
so the terminal mode (-t option) got changed to use it. No longer
scanning all the device files is required, we just look for processes
that have a tty and group them together.

This is a much neater method and also means we get output even
when we are using utmp on a system with no utmp, with some
fields missing.

References:
 commit f53cc24d57085c87ebb1871b92c0069b72a60926

Signed-off-by: Craig Small <csmall@dropbear.xyz>