watch: Add --follow option
Sometimes you want to keep the output from watch on the screen instead
of clearing each command run. The --follow option will scroll the
output window continuosly.  Any option that needs to know what
is printed doesn't work with it so watch prints an error if
you try.
watch now is two windows in a screen; a 2 row header and a
remainder for the output.

Signed-off-by: Craig Small <csmall@dropbear.xyz>