watch: 256 color support
After an unhealthy amount of research into how ncurses does its thing
and the sad tale of ANSI color sequences I think the 256 colors for
watch nut have been cracked.

Because why would you set your color pair in attrs or pair when
(void*) opts will do? Also the colors are defined on a 0-1000
scale because... for reasons beyond the knowledge of all, including
Thomas Dickey.

The pair definitions needed to be changed to their extended versions
as the short values won't hold all the combinations.

The SGR sequence, which uses the E[38;5;X is now done for all 256
colors. I'm not sure if 24-bit (E[32;2;r;g;b]) will ever be done.
Why would anyone need 24-bit command line tools?!

I'm not even sure what a E]4;X; sequence is so that's not
likely to be supported either.

References:
 #44
 https://en.wikipedia.org/wiki/ANSI_escape_code#8-bit
 https://invisible-island.net/ncurses/man/curs_attr.3x.html
 https://linux.die.net/man/3/wcolor_set
 https://stackoverflow.com/questions/43710131/how-does-ncurses-init-color-function-translate-to-traditional-rbg-colors/43710790
 https://gist.github.com/hSATAC/1095100

Signed-off-by: Craig Small <csmall@dropbear.xyz>