watch: Extended functions in wide ncurses
The referenced commit added 256 colors to watch using the extended
versions of init_color/pair. While these appear in all modern versions
of the wide ncurses library, they don't in non-wide versions.

While it is not a 100% overlap between a ncurses library that supports
wide characters and a ncurses library that support extended colors, the
intersection is minor; just enable watch 8 bit with
./configure --enable-watch8bit to get wide and colorful displays.

References:
 commit 0349d5fa37592d0f48633ef0f7c6c8e4709b928b
 https://man7.org/linux/man-pages/man3/curs_color.3x.html

Signed-off-by: Craig Small <csmall@dropbear.xyz>