pgrep: Match on process ID
While it might seem initially odd to use pgrep to
kill a process based upon the PID when kill does this
faster, it gives the opportunity to use PID and some other
criteria, e.g. match all the PIDS that have this user.

The use-cases for pidfile and pid options are different
and it would have been a fiddle to enable both at the
same time so the easiest way is to conflict.

References:
 https://bugs.debian.org/612146

Signed-off-by: Craig Small <csmall@dropbear.xyz>