sysctl: Use options after --system
If sysctl saw the --system option it would immediately process the
system files and exit, ignoring anything on the command line after
it.

This would mean, for example, if you wanted to use the --pattern
option to filter out keys it would work if --pattern was before
--system but not the other way around.

I made a slight change from Simon's suggestion to process
keys on the command line, so something like
sysctl --system foo=bar
also works, it seems like an odd way of running sysctl but it
is valid in a way.

References:
 https://bugs.debian.org/978989

Signed-off-by: Craig Small <csmall@dropbear.xyz>