ps: Make parser understand wide helps
ps is able to understand the --help option in a translated
language. For the parser, if the first character of the translated
word "help" was an ASCII letter, the parser could understand it.
For other languages such as Korean which use wide characters this
check failed. The parser was updated to check for the next
wide character after -- is an alpha wide character.

References:
 #381

Signed-off-by: Craig Small <csmall@dropbear.xyz>