ps: Add a column delimiter
With the --delimter option you can set a column delimiter instead
of variable space:
$ src/ps/pscommand --delimiter '|'
PID|TTY|TIME|CMD
33931|pts/1|00:00:00|bash
361622|pts/1|00:00:00|pscommand

It doesn't escape the fields, so if colon was used for the previous
command then the colons in the time and the delimiter would be mixed
up.

Other than parser and error checking, it was a two-line difference,
the future will tell if there should have been more than two.

References:
 #118

Signed-off-by: Craig Small <csmall@dropbear.xyz>