ps: Fix forest mode
Initially the issue said that while ps -j showed a forest, ps f would
not in a docker container. Further investigation showed it was
everywhere for direct decendents of the init process and it was
there at least since procps 3.3.12

The offending code was:
    if(self_pid==1 && ADOPTED(processes[i]) && forest_type!='u')
      show_tree(i++, n, level, more_children);
    else
      show_tree(i++, n, IS_LEVEL_SAFE(level+1) ? level+1 : level, more_children);

option -j is forest_type u while option f is option type b.
ADOPTED is always true; so if the parent is pid=1 and we didn't
use the -j variant then the tree level of the children was 0, or
the top of the tree.

This looks like it's been there since 2002 as an import of
procps 010114 shows this line.

At the same time, I cleaned up the function to make it easier
to read.

References:
 #102
 commit 03a9b5a30fb773ee4d1568064b14e075ba186c90

Signed-off-by: Craig Small <csmall@dropbear.xyz>