pgrep: use pidfd_send_signal
When you use kill or sigqueue there is a possibility that the PID
is reused during the run of pgrep.

Using pidfd_open locks the file descriptor to the process that
was using that pid at the time of the open. Then all subsequent
checks or operations are on the same process.

Updated pgrep.1 to inform users how the signal is sent and
some cleanups:
 Ordered the options alphabetically
 Removed the out of date and discouraged (by man-pages(7) Authors section
References:
 #207

Signed-off-by: Craig Small <csmall@dropbear.xyz>