w: Use correct return value for sd_get_sessions
If procps is compiled with systemd support but run on a non-systemd
host, while the functions in sd_get_sessions get ENOENT (because the
directory doesn't exist) the function returns EPERM instead; which
is strictly not true but we will check for both.

The man page is silent on what it should be.

References:
 https://bugs.debian.org/10698904
 https://manpages.debian.org/stretch-backports/libsystemd-dev/sd_get_sessions.3.en.html

Signed-off-by: Craig Small <csmall@dropbear.xyz>