ps: Better error messages for simple_spew
The simple_spew function would print an error message if there was an
issue with the library. The issue was:
 * The same message was used for a _reap or _select which mentioned
   reap only.
 * We generally use x?err() function for fatal errors, which also
   gives the errno of the failed function

Both these issues meant debugging what was going on took longer than
it should. A problem inside _select and the code path leading to it
would not be considered because the error message is about _reap.
With no errno, knowing where the library function failed and returned
is unknown.

This change will mean problems such as #387 will get diagnosed quicker.

References:
 #387

Signed-off-by: Craig Small <csmall@animal.local>