top: duck '-Wunused-result, -Wempty-body' gcc warnings
The program source implied that the original construct
was to avoid an '-Wunused-result' warning, but I don't
believe it was ever a potential. In any case, the code
produced the latest '-Wempty-body' warning, suggesting
braces be employed in place of that current semicolon.

This commit will employ a slightly different approach.

[ warning was enabled with that '-Wextra' gcc option ]

Reference(s):
warning: suggest braces around empty body in an 'if' statement

Signed-off-by: Jim Warner <james.warner@comcast.net>