top: refactored and expanded secondary help processing
The primary top help screen had grown way too crowded.

In fact there was no room left to represent the Ctrl-A
command, which was recently added to display permitted
capabilities as part of the 'bottom window' provision.

So, this commit will now represent all 'bottom window'
commands on their own separate screen which is reached
by keying 'h' or '?' while on the primary help screen.

Signed-off-by: Jim Warner <james.warner@comcast.net>