library: address a potential race resulting in a fault
The library position was to never subject callers to a
NULL string pointer. Rather, a literal "-" was usually
substituted to mean such information is not available.

But, as revealed in the issue cited below, there was a
huge unknown gap in our goal. Under the original code,
if the 'status' file was not available in /proc/<pid>,
fields like supgrp, ruser, rgroup, etc. would be NULL.

The same was true with lxc and docker container fields
which depend on the 'cgroup' file. Should that file be
missing it means more unexpected NULL string pointers.

Under what conditions might such files be unavailable?
If a process exited but has not yet been fully reaped,
it could be encountered by a repeatedly run ps or by a
continuously running program like top. In such a case,
there's a possibility some files under /proc/<pid> are
missing. A NULL string address could then be a result.

[ as workload increases on a machine with many cores ]
[ that risk of a missing file rises. the environment ]
[ prompting this issue involved a large Epyc machine ]
[ running 72 concurrent testing processes & invoking ]
[ the ps program on average five times every second. ]

So, this commit will initialize the 'status' dependent
string results with a pointer to a "-" string variable
whose address can be known. This will save us the cost
of strdup() calls. Then, that pids module can test for
it to avoid any free() attempt when assigning results.

[ for lxc & docker containers, that require 'cgroup' ]
[ to be present, the solution to some potential NULL ]
[ pointer is addressed in a little different manner. ]

Lastly, the 'stat' file is now read unconditionally to
prevent a NULL 'cmdline' pointer for kernel threads if
the 'status' file is missing or wasn't otherwise read.

Many thanks to Mike Gulick for raising this issue plus
Chao Liu for his contributions. Also, a special thanks
to Tsukasa OI for an analysis leading to the solution.

Reference(s):
https://gitlab.com/procps-ng/procps/-/issues/380

Signed-off-by: Jim Warner <james.warner@comcast.net>