library: avoid much code if LIBPROC_HIDE_KERNEL active
Since that hide kernel threads feature was introduced,
code deciding to keep or discard a proc_t was found at
the very end of those functions valuing the structure.

On the surface this may seem reasonable since that was
where the actual 'return p' statement was to be found.

But in reality it made no sense to work through all of
those PROC_FILL, etc flags only to toss the results if
it turns out some kernel thread was not really wanted.

So, this commit just relocates the appropriate code to
the earliest possible point thus saving that overhead.

[ and now that a 'stat' file is read unconditionally ]
[ a ppid is guaranteed early, like a tid already is, ]
[ so some special code under openproc() isn't needed ]

Reference(s):
. Jan, 2022 - original hide_kernel implementation
commit 2a7ec67ac806d91520f59c384684e2332774368b

Signed-off-by: Jim Warner <james.warner@comcast.net>