library: fix another potential race, this one harmless
Occasionally I'd encounter that valgrind warning shown
below when testing the top program. It only appears if
thread mode is active, and even so was extremely rare.

Ultimately, it is a function of the delay interval and
the child threads lifetime. With faster iterations and
or shorter lived threads, the warning was more likely.

It was caused by a failed open plus a -1 in PT->pidfd.

Between the time the next process is identified in the
simple_nextpid() function via readdir and the time the
process' /proc/# directory is opened it may have died.
So a failed open went undetected yielding the negative
PT->pidfd field. Then, readeither() passed the invalid
descriptor to simple_nexttid() who called opendirat().
And that's where the mistake was detected by valgrind.

The bottom line is that unlike our earlier recent race
producing the potential segfault, this one just delays
slightly that decision to move on to the next process.

[ but that is no excuse to not apply this simple fix ]
[ and in the process make peace with valgrind again! ]

Reference(s):
Warning: invalid file descriptor -1 in syscall openat()

Signed-off-by: Jim Warner <james.warner@comcast.net>