library: avoid possible segmentation fault, <pids> api
In the issue referenced below, that SEGV was caused by
passing a '-1' pid to the procps_pids_select function.
That value, in turn, is caused by providing a LONG_MAX
or greater number for what was assigned to an integer.

So, this commit will help ensure we avoid a match when
we should fail plus skip some unproductive libc calls.

[ the same problem occurs in that old 3.3.17 library ]

Reference(s):
https://gitlab.com/procps-ng/procps/-/issues/394

Signed-off-by: Jim Warner <james.warner@comcast.net>