top: add feature to display physical cores versus cpus
The program has long been able to display cpu and numa
nodes. More recently it can even show the type of core
a particular cpu is associated with (assuming a proper
Intel based platform). Now, with this patch it will be
able to also show those physical core statistics as an
alternative to the individual or combined cpu display.

Signed-off-by: Jim Warner <james.warner@comcast.net>